edited rating type and win/lose resolvers to update rating

+++ server/schemas/typeDefs.js
@@ -8,7 +8,7 @@ const typeDefs = gql`
         wins: Int!
         losses: Int!
         gamesPlayed: Int!
-        rating: Int!
+        rating: Float!
         team: Team
         createdAt: String
         updatedAt: String
